pyproject.toml
[project]
name = "axonium"
version = "0.5.0"
description = "SDK para consumir llama-server con autenticación y renovación de tokens"
readme = "README.md"
requires-python = ">=3.13"
authors = [
  { name = "Emeric Espiritu Santiago" }
]
dependencies = [
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "openai>=2.15.0",
  "pydantic>=2.5.0",
  "langchain-openai>=0.1.0",
  "langchain-core>=0.1.0",
  "langfuse>=2.0.0",
  "llm-guard>=0.3.16",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
